Make things prettier
Pass down if we are on a twitch/youtube server to the buffer renderer so
we can add space for badges even if the messages dont have badge tags.

Make the timestamp + username a span so that only the message is valid
for line wrapping breaks.

Adjust style for the above and nicer emote alignment.